requirements.txt
PyQt6>=6.5.0
PyMuPDF>=1.23.0
Pillow>=10.0.0
pyinstaller>=6.0.0
